pyproject.toml
[project]
name = "agencli-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-frontmatter",
    "langchain",
    "langchain-openai",
    "python-dotenv",
    "rich"
]


